cross_cis_scanner_package/requirements.txt
pywinrm
paramiko
pyyaml
